Reduce duplication in npm-related tasks
